fix: 修复dependsInstall命令注入漏洞
增加对传入包路径的判断。

Log:  增加对传入包路径的判断。
Bug: https://pms.uniontech.com/zentao/bug-view-72431.html

+++ src/AptInstallDepend/installDebThread.cpp
@@ -41,7 +41,7 @@ void InstallDebThread::setParam(QStringList tParam)
 
 void InstallDebThread::getDescription()
 {
-    QString str = "sudo dpkg -e " + m_listParam[1] + " " + TEMPLATE_DIR;
+    QString str = "dpkg -e " + m_listParam[1] + " " + TEMPLATE_DIR;
     system(str.toUtf8());
 
     QFile file;
@@ -114,21 +114,26 @@ void InstallDebThread::run()
         } else if (m_listParam[0] == "InstallConfig") {
             if (m_listParam.size() <= 1)
                 return;
+            const QFileInfo info(m_listParam[1]);
+            const QFile debFile(m_listParam[1]);
 
-            qDebug() << "StartInstallAptConfig";
+            if (debFile.exists() && info.isFile() && info.suffix().toLower() == "deb") {        //大小写不敏感的判断是否为deb后缀
+                qDebug() << "StartInstallAptConfig";
 
-            getDescription();
+                getDescription();
 
-            //m_proc->start("sudo", QStringList() << "-S" <<  "dpkg-preconfigure" << "-f" << "Teletype" << m_listParam[1]);
-            m_proc->start("sudo", QStringList() << "-S" <<  "dpkg" << "-i" << m_listParam[1]);
-            m_proc->waitForFinished(-1);
+                //m_proc->start("sudo", QStringList() << "-S" <<  "dpkg-preconfigure" << "-f" << "Teletype" << m_listParam[1]);
+                m_proc->start("sudo", QStringList() << "-S" <<  "dpkg" << "-i" << m_listParam[1]);
+                m_proc->waitForFinished(-1);
 
-            QDir filePath(TEMPLATE_DIR);
-            if (filePath.exists()) {
+                QDir filePath(TEMPLATE_DIR);
+                if (filePath.exists()) {
                 filePath.removeRecursively();
-            }
+                }
 
-            m_proc->close();
+                m_proc->close();
+                
+            }
         }
     }
 }
